reproject: pick integer nodata sentinel for int rasters (#2185) (#2191)
* reproject: pick integer nodata sentinel instead of NaN for int dtypes (#2185)

Integer rasters reprojected without an explicit nodata used to lose
out-of-bounds pixels to silent 0s while attrs['nodata'] kept claiming
NaN. The worker rounds and casts back to the input integer dtype, and
NaN doesn't survive that cast. Pass the raster dtype into
_detect_nodata so integer inputs get a dtype-appropriate sentinel
(dtype.min for signed, dtype.max for unsigned) following the
rasterio/GDAL convention.

The same fix applies to merge's per-input nodata detection so mixed
integer inputs canonicalize correctly during the mosaic step.

* Address review: swap NaN nodata for int sentinel post-resolution (#2185)

- `_detect_nodata` now applies the dtype-aware swap after resolving
  the raw value, so explicit `attrs['nodata']=nan` or
  `nodata=float('nan')` on an integer raster also yields the integer
  sentinel. The split-out `_detect_nodata_raw` keeps the lookup chain
  readable.
- Add tests for the NaN-in-attrs and explicit-NaN-arg cases.
- Rename `TestReprojectIntegerNodataNumpy` to flag the class-level
  parametrize, and note that future contributors should not add
  non-parametric methods to it.
- Expand `_int_raster_with_oob` docstring to call out the EPSG:32633
  dependence.

+++ xrspatial/reproject/__init__.py
@@ -662,8 +662,12 @@ def reproject(
         )
     tgt_crs = _resolve_crs(target_crs)
 
-    # Detect nodata
-    nd = _detect_nodata(raster, nodata)
+    # Detect nodata. Pass the raster dtype so integer rasters get an
+    # integer-compatible sentinel (dtype min for signed, dtype max for
+    # unsigned) instead of NaN. Without this hint, the worker's
+    # cast-back step would collapse NaN to 0 and `attrs['nodata']`
+    # would contradict the array contents (#2185).
+    nd = _detect_nodata(raster, nodata, dtype=raster.dtype)
 
     # Source geometry
     src_bounds = _source_bounds(raster)
@@ -1769,7 +1773,11 @@ def merge(
             "Could not detect target CRS. Pass target_crs explicitly."
         )
 
-    # Detect output nodata (the sentinel the user asked for)
+    # Detect output nodata (the sentinel the user asked for). The merge
+    # output is always float64, so NaN is fine as the output sentinel
+    # when the user didn't supply one -- the integer-cast hazard from
+    # #2185 only applies to the per-input ``r_nd`` values that flow into
+    # the per-raster reproject worker.
     nd = nodata if nodata is not None else _detect_nodata(rasters[0], nodata)
 
     # Gather source info for each raster
@@ -1787,8 +1795,10 @@ def merge(
         yd = _is_y_descending(r)
         # Per-raster input nodata sentinel. Detected independently of the
         # user-supplied output nodata so that mixed-sentinel inputs are
-        # canonicalized correctly during merge.
-        r_nd = _detect_nodata(r, None)
+        # canonicalized correctly during merge. Pass the raster dtype so
+        # integer sources get an integer-compatible sentinel rather than
+        # NaN -- the same fix as the reproject() path for #2185.
+        r_nd = _detect_nodata(r, None, dtype=r.dtype)
         raster_infos.append({
             'raster': r,
             'src_crs': src_crs,
